# openapi.yaml
openapi: 3.1.0
info:
  title: ClassnoteX API
  description: Backend API for ClassnoteX - AI-powered recording and transcription
    app.
  version: 0.4.0
paths:
...
  /usage/admin/backfill:
    post:
      tags:
      - Usage
      - Usage
      summary: Backfill Usage
      description: 'Admin-only: Backfill usage data from sessions collection.

        Scans all sessions and populates user_daily_usage.'
      operationId: backfill_usage_usage_admin_backfill_post
      parameters:
      - name: secret
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Secret
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /usage/me/summary:
    get:
      tags:
      - Usage
      - Usage
      summary: Get My Usage Summary
      description: 'Get the current user''s usage summary for a date range.


        Useful for:

        - Showing usage dashboard in the app

        - Checking quota usage

        - Billing purposes'
      operationId: get_my_usage_summary_usage_me_summary_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: from_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Start date (yyyy-MM-dd). Defaults to 30 days ago.
          title: From Date
        description: Start date (yyyy-MM-dd). Defaults to 30 days ago.
      - name: to_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: End date (yyyy-MM-dd). Defaults to today.
          title: To Date
        description: End date (yyyy-MM-dd). Defaults to today.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageSummaryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /usage/admin/users/{user_id}/summary:
    get:
      tags:
      - Usage
      - Usage
      summary: Get User Usage Summary Admin
      description: 'Admin endpoint to view any user''s usage.


        TODO: Add proper admin role check'
      operationId: get_user_usage_summary_admin_usage_admin_users__user_id__summary_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          title: User Id
      - name: from_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: From Date
      - name: to_date
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: To Date
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageSummaryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /usage/analytics/me/timeline:
    get:
      tags:
      - Usage
      - Usage
      summary: Get Usage Timeline
      description: 時系列データの取得（グラフ用）。
      operationId: get_usage_timeline_usage_analytics_me_timeline_get
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: range
        in: query
        required: false
        schema:
          type: string
          default: 30d
          title: Range
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    UsageSummaryResponse:
      properties:
        user_id:
          type: string
          title: User Id
        from_date:
          type: string
          title: From Date
        to_date:
          type: string
          title: To Date
        total_recording_sec:
          type: number
          title: Total Recording Sec
          default: 0.0
        session_count:
          type: integer
          title: Session Count
          default: 0
        summary_invocations:
          type: integer
          title: Summary Invocations
          default: 0
        summary_success:
          type: integer
          title: Summary Success
          default: 0
        quiz_invocations:
          type: integer
          title: Quiz Invocations
          default: 0
        quiz_success:
          type: integer
          title: Quiz Success
          default: 0
        diarization_invocations:
          type: integer
          title: Diarization Invocations
          default: 0
        qa_invocations:
          type: integer
          title: Qa Invocations
          default: 0
        llm_input_tokens:
          type: integer
          title: Llm Input Tokens
          default: 0
        llm_output_tokens:
          type: integer
          title: Llm Output Tokens
          default: 0
        share_count:
          type: integer
          title: Share Count
          default: 0
        export_count:
          type: integer
          title: Export Count
          default: 0
        passRate:
          anyOf:
          - type: number
          - type: 'null'
          title: Passrate
        topTags:
          anyOf:
          - items: {}
            type: array
          - type: 'null'
          title: Toptags
        timeline_daily:
          anyOf:
          - items: {}
            type: array
          - type: 'null'
          title: Timeline Daily
        by_mode:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: By Mode
      type: object
      required:
      - user_id
      - from_date
      - to_date
      title: UsageSummaryResponse
      description: Response for /me/usage-summary
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: token